Converted doubles to floats.

+++ src/main/java/edu/emory/mathcs/nlp/vsm/optimizer/HierarchicalSoftmax.java
@@ -32,7 +32,7 @@ public class HierarchicalSoftmax extends Optimizer
 	}
 	
 	@Override
-	public void learnBagOfWords(Random rand, int word, float[] syn1, float[] neu1, float[] neu1e, double alpha)
+	public void learnBagOfWords(Random rand, int word, float[] syn1, float[] neu1, float[] neu1e, float alpha)
 	{
 		byte[] code  = vocab.get(word).code;
 		int [] point = vocab.get(word).point;
@@ -42,7 +42,7 @@ public class HierarchicalSoftmax extends Optimizer
 	}
 	
 	@Override
-	public void learnSkipGram(Random rand, int word, float[] syn0, float[] syn1, float[] neu1e, double alpha, int l1)
+	public void learnSkipGram(Random rand, int word, float[] syn0, float[] syn1, float[] neu1e, float alpha, int l1)
 	{
 		byte[] code  = vocab.get(word).code;
 		int[]  point = vocab.get(word).point;
